Fixed loss of precision on length tools

+++ DrawCircleMode.cs
@@ -291,9 +291,10 @@ namespace TriDelta.DrawCircleMode {
          if (handleInner == null || editSides < 3)
             return;
 
-         double aoe = (360 / editSides) / 2;
+         double aoe = (360 / (float)editSides) / 2;
          double rads = aoe * (Math.PI / 180);
          float length = (float)((newlength / 2) / Math.Sin(rads)); //assume a right angle (sin(90) = 1)
+
          SetLength(length, true);
          UpdateLengthBox();
       }
